feat:fixing preAuthorize in controller
